basic enemy ai selecting skill based on likelihood

+++ Assets/Scripts/Data/ScriptableObjects/SkillData/Skill.cs
@@ -22,5 +22,5 @@ public class Skill : ScriptableObject
     public Animation TargetAnimation;
     public ActionTypes ActionType;
     public int Price;
-    public float Likelihood;
+    public int Likelihood;
 }
